% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{add.effect}
\alias{add.effect}
\title{find the QTL addtive and dominant effect for all traits in the transband.}
\usage{
add.effect(transbands, cross)
}
\description{
find the QTL addtive and dominant effect for all traits in the transband.
}

