% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/add.geno.R
\name{add.geno}
\alias{add.geno}
\title{find the peak of scanone.mvn and the QTL genotype.}
\usage{
add.geno(transbands, cross, regn.cM = 5, max.p = 100)
}
\arguments{
\item{transbands}{result from find.trans, a list of transband.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{regn.cM}{disctance in cM form QTL for defining non-recombinanct mice.}

\item{max.p}{max number of expression traits used.}
}
\description{
For a list of hotspots, find the single QTL genotype and the mice
with no recombinant events in a 10cM region around the single QTL.
}

