% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/add.pheno.R
\name{add.pheno}
\alias{add.pheno}
\title{for each transband, add the phenotypes "Y".}
\usage{
add.pheno(transbands, mlratio)
}
\arguments{
\item{transbands}{result from find.trans, a list of transband.}

\item{mlratio}{matrix of expression traits}
}
\description{
for each transband, add the phenotypes "Y".
}

