% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{fit a classification model with "LDA", "KNN" or "SVM"}
\usage{
classification(data.train, data.test, class.train, method = c("LDA", "KNN",
  "SVM"), K = 50)
}
\arguments{
\item{data.train}{traing data that is used to build the classifier.}

\item{data.test}{testing data that need to be classified.}

\item{class.train}{Label of training data.}

\item{method}{"LDA", "KNN" or "SVM", default is "LDA".}

\item{K}{parameter for k-nearest neighbor method.}
}
\value{
a list with pred.test, pred.score, pred.train and
error.train. when method is "LDA", will also return the scaling
vector "sca" that could be used to generate the Linear
Discriminants.
}
\description{
fit a classification model with "LDA", "KNN" or "SVM"
}

