% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/convert_scan1.R
\name{convert_scan1}
\alias{convert_scan1}
\title{Summarize result of scanone}
\usage{
convert_scan1(out, phenoname, chr = 1:19)
}
\arguments{
\item{out}{output from R/scanone}

\item{phenoname}{names of the traits of interest}

\item{chr}{chromosomes  of interest}
}
\value{
data frame with \code{pheno, chr, pos}, and \code{lod1}, a
summary of scan1 result, max LOD score and its position for each
phenoname and chromosome combinantion.
}
\description{
Turn the result of scanone to a smaller and thinner dataframe with
max LOD score and its position for each trait and each chrosome.
}

