% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{count.trans}
\alias{count.trans}
\title{From a scan1 result, count how many trans-eQTL are there in every
bin across the chromosomes.}
\usage{
count.trans(out1, probepos, chr, marker.info, lod.thr = 5, trans.cM = 5,
  kernal.width = 1, window.cM = 10)
}
\arguments{
\item{out1}{summary of scan1 result}

\item{probepos}{Genetic position for each probe in cM.}

\item{chr}{Optional vector indicating the chromosomes considered.}

\item{marker.info}{Data.frame of marker information, result of
\code{get.marker.info}}

\item{lod.thr}{LOD threshold}

\item{trans.cM}{threshold of the distance between the qtl position
and the probe position to call a eQTL cis.}

\item{kernal.width}{for weighted count, kernal.width is used as sd
value to generate normal weights. These weights are then used as
kernal function to sum over each bin/window.}

\item{window.cM}{bin width in cM.}
}
\value{
a data.frame with "chr", "pos", "count", "count.sum" and
"count.wtsum".
}
\description{
From a scan1 result, count how many trans-eQTL are there in every
bin across the chromosomes.
}

