% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{det_AtA}
\alias{det_AtA}
\title{Determinant of self cross product}
\usage{
det_AtA(X, logarithm = TRUE)
}
\arguments{
\item{X}{matrix}

\item{logarithm}{A bool value indicating whether determinant or
log determinate should be returned.}
}
\value{
The determinant or log determinant
}
\description{
Calculate the determinate of X^T X. By using the symmetric property
of X^T X, this implementation is much faster than calling the
general function `determinant` on crossproduct(X)
}

