% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/qtlpvl-package.R
\docType{data}
\name{fake.probepos}
\alias{fake.probepos}
\title{Simulated Genetic Positions for Gene Expression Data Set}
\format{A data frame with 10 rows and 2 columns: 'chr' and 'cM'.}
\description{
Simulated position for the 10 gene expression traits with
chromosome number and genetic positions in cM.
}
\details{
These data were simulated using the genotype data from
the `listeria` data set (an F$_2$ population) provided with
R/qtl. The phenotypes were simulated using two markers from
chromosome 1 as QTL, with the first QTL having an additive allelic
effect, and with one of the alleles at the second QTL being
strictly dominant. There are 10 phenotypes. The first 5 are
controlled by the first QTL, and the other 5 traits are controlled
by the second QTL (and with a negative and larger effect). The 10
phenotypes were generated with these QTL effects plus independent,
normally distributed residual variation. Treating these traits as
gene expression measurements, we assigned genomic positions at
random. The phenotype data is stored in matrix `fake.phenos` and their
positions are stored in data frame `fake.probepos`.
}
\keyword{datasets}

