% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/find.trans.R
\name{find.trans}
\alias{find.trans}
\title{From a scan1 result, first count how many trans-eQTL are there in
every bin across the chromosomes, then find phenos inside each
transband.}
\usage{
find.trans(out1, probepos, chr, marker.info, lod.thr = 5, trans.cM = 5,
  kernal.width = 1, window.cM = 10, trans.count.thr = 300)
}
\arguments{
\item{out1}{summary of scan1 result}

\item{probepos}{Genetic position for each probe in cM.}

\item{chr}{Optional vector indicating the chromosomes considered.}

\item{marker.info}{Data.frame of marker information, result of
\code{get.marker.info}}

\item{lod.thr}{LOD threshold}

\item{trans.cM}{threshold of the distance between the qtl position
and the probe position to call a eQTL cis.}

\item{kernal.width}{for weighted count, kernal.width is used as sd
value to generate normal weights. These weights are then used as
kernal function to sum over each bin/window.}

\item{window.cM}{bin width in cM.}

\item{trans.count.thr}{threshold used to define a region as
transband.}
}
\value{
a list of transband.
}
\description{
From a scan1 result, first count how many trans-eQTL are there in
every bin across the chromosomes, then find phenos inside each
transband.
}

