% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get.marker.info.R
\name{get.marker.info}
\alias{get.marker.info}
\title{Marker information.}
\usage{
get.marker.info(cross, chr)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{chr}{Optional vector indicating the chromosomes for which
markers info need to be pulled out. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}
}
\value{
A data.frame whose first column contains the chromosome
IDs, second column contains cM positions, row names are
markernames.
}
\description{
Pull out marker names positions and chromosomes as a data frame.
}
\examples{
data(listeria)
marker <- get.marker.info(listeria, chr=1:19)
}

