% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get.trans.info.R
\name{get.trans.info}
\alias{get.trans.info}
\title{For a scan1 result and probepos, find if the eQTLs are cis-eQTL or trans-eQTL.}
\usage{
get.trans.info(out1, probepos, chr = NULL, marker.info, lod.thr = 5,
  trans.cM = 10)
}
\arguments{
\item{out1}{summary of scan1 result}

\item{probepos}{Genetic position for each probe in cM.}

\item{chr}{Optional vector indicating the chromosomes considered.}

\item{marker.info}{Data.frame of marker information, result of
\code{get.marker.info}}

\item{lod.thr}{LOD threshold}

\item{trans.cM}{threshold of the distance between the qtl position
and the probe position to call a eQTL cis.}
}
\value{
input data.frame with two added columns, "is.cis" and
"is.trans".
}
\description{
For a scan1 result and probepos, find if the eQTLs are cis-eQTL or trans-eQTL.
}

