% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{group.train.test}
\alias{group.train.test}
\title{find the training and tesing set from genotypes in a given chr region.}
\usage{
group.train.test(cross, Y, chr, region.l, region.r)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{traits mapped to the region.}

\item{chr}{which chromosome is under invest.}

\item{region.l}{left bound of region under invest.}

\item{region.r}{right bound of region under invest.}
}
\description{
find the training and tesing set from genotypes in a given chr region.
}
\details{
...
}

