% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{interact}
\alias{interact}
\title{computes a matrix represents the interaction of two given matrices.}
\usage{
interact(x, y)
}
\arguments{
\item{x}{Matrix}

\item{y}{Matrix}
}
\value{
Matrix of interaction of x and y
}
\description{
computes a matrix represents the interaction of two given matrices.
}

