% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{find interval for multiple traits mapped to the same region.}
\usage{
interval(cross, Y, chr, region.l, region.r, do.plot = TRUE, main = "",
  weighted = FALSE, thr = 0.9)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{traits mapped to the region.}

\item{chr}{which chromosome is under invest.}

\item{region.l}{left bound of region under invest.}

\item{region.r}{right bound of region under invest.}

\item{do.plot}{logical, plot results? default is \code{TRUE}}

\item{main}{main titile of the plot.}

\item{weighted}{use pred.score as weights? default is \code{FALSE}.}

\item{thr}{cutoff for the interval. default is 0.9}
}
\value{
list of chr and interval and plots if \code{do.plot=TRUE}.
}
\description{
find interval for multiple traits mapped to the same region.
}

