% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm.resid}
\alias{lm.resid}
\title{Calculte residual matrix.}
\usage{
lm.resid(X, Y, method = c("llt", "qr", "svd"))
}
\arguments{
\item{X}{a model matrix}

\item{Y}{thr response matrix}

\item{method}{'llt' for the LLT Cholesky, 'qr' for the
column-pivoted QR decomposition, 'svd' for the Jacobi singular
value decomposition (SVD)}
}
\value{
residual matrix.
}
\description{
Run lm(Y~X) with specific method through Rcpp and return a
residual matrix.
}

