% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_resid_symmEigen}
\alias{lm_resid_symmEigen}
\title{Calculate the residual of lm(Y ~ X)}
\usage{
lm_resid_symmEigen(X, Y, threshold = 1e-07)
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}

\item{threshold}{Eigen decomposition is used in calculation. An
eigen value smaller than threshold * largest eigen value is
considered zero.}
}
\value{
The residual matrix
}
\description{
Calculate the residual of lm(Y ~ X)
}

