% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lmresid_svd}
\alias{lmresid_svd}
\title{Residual of lm(Y ~ X)}
\usage{
lmresid_svd(X, Y, threshold = 1e-07)
}
\arguments{
  \item{X}{matrix}

  \item{Y}{matrix}

  \item{threshold}{Eigen decomposition is used in
  calculation. An eigen value smaller than threshold *
  largest eigen value is considered zero.}
}
\value{
The residual matrix
}
\description{
An implementation for calculate residual of linear
regression. It's ' not as fast as lmresid_llt (about 50% slower),
but is rank revealing, so should be safe for linear correlated X.
}

