% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/make.transbands.R
\name{make.transbands}
\alias{make.transbands}
\title{From a scan1 result, first count how many trans-eQTL are there in
every bin across the chromosomes, then find phenos inside each
transband. Add all pheno and geno infomations needed.}
\usage{
make.transbands(out1, probepos, cross, chr, marker.info, mlratio, lod.thr = 5,
  trans.cM = 5, kernal.width = 1, window.cM = 10, trans.count.thr = 300,
  regn.cM = 5)
}
\arguments{
\item{out1}{summary of scan1 result}

\item{probepos}{Genetic position for each probe in cM.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{chr}{Optional vector indicating the chromosomes considered.}

\item{marker.info}{Data.frame of marker information, result of
\code{get.marker.info}}

\item{lod.thr}{LOD threshold}

\item{trans.cM}{threshold of the distance between the qtl position
and the probe position to call a eQTL cis.}

\item{kernal.width}{for weighted count, kernal.width is used as sd
value to generate normal weights. These weights are then used as
kernal function to sum over each bin/window.}

\item{window.cM}{bin width in cM.}

\item{trans.count.thr}{threshold used to define a region as
transband.}

\item{regn.cM}{disctance in cM form QTL for defining non-recombinanct mice.}
}
\value{
a list of transband.
}
\description{
From a scan1 result, first count how many trans-eQTL are there in
every bin across the chromosomes, then find phenos inside each
transband. Add all pheno and geno infomations needed.
}

