% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ploteQTL.R
\name{ploteQTL}
\alias{ploteQTL}
\title{eQTL plot}
\usage{
ploteQTL(marker.info, probepos, phenonames, markers, chr1, pos1, main = "",
  plot.chr = c(1:19, "X"), add = FALSE, col = "black")
}
\arguments{
\item{marker.info}{information of markers}

\item{probepos}{probe position}

\item{phenonames}{names of probes to be ploted}

\item{markers}{marker names of eQTL}

\item{chr1}{chr of eQTLs}

\item{pos1}{pos of eQTLs}

\item{main}{main title of plot}

\item{plot.chr}{chrs where probes from to be ploted}

\item{add}{start a new plot or add to an existed one.}

\item{col}{color of dots}
}
\description{
eQTL plot
}

