% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{plotlda}
\alias{plotlda}
\title{plot of ld2 vs ld1.}
\usage{
plotlda(data.train, data.test, class.train, pred.test, error.train, sca,
  main = "")
}
\arguments{
\item{error.train}{LDA traning error}

\item{sca}{scalling factor for ld1 and ld2}

\item{pred.test}{vector of length n, predicted class (1, 2 or 3)
for each sample.}

\item{main}{main titile of the plot.}

\item{data.train}{traing data that is used to build the classifier.}

\item{data.test}{testing data that need to be classified.}

\item{class.train}{Label of training data.}
}
\value{
plot of ld2 vs ld1
}
\description{
plot training and testing data together, to compare their
pattern. different pattern indicates multiple QTLs, because it
needs extra signal to drag the tesing points away from the traing
points.
}

