% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotLOD.R
\name{plotLOD}
\alias{plotLOD}
\title{Plot LOD score versus QTL position for multiple traits.}
\usage{
plotLOD(Y, cross, chr, maxLOD, maxPOS, addcovar = NULL, intcovar = NULL,
  LOD.threshold = 3, ...)
}
\arguments{
\item{Y}{Matrix of multiple traits, with samples in the row,
traits in the column.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{maxLOD}{max LOD score for each trait}

\item{maxPOS}{position of maxLOD for each trait}

\item{addcovar}{Additive covariates, samples in the row.}

\item{intcovar}{Interactive covariates (interact with QTL genotype).}

\item{LOD.threshold}{threshold for QTL to be displayed.}

\item{...}{Optional graphics arguments}
}
\value{
A plot LOD score versus QTL position for multiple traits.
}
\description{
For each trait, plot its LOD score versus QTL position when the
LOD is bigger than a threshold.
}
\examples{
data(fake.phenos)
data(listeria)
listeria <- calc.genoprob(listeria, step=1)
chr <- 1
plotLOD(Y, listeria, chr)
}

