% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plottrace.R
\name{plottrace.testpleio.1vs2}
\alias{plottrace.testpleio.1vs2}
\title{plot trace}
\usage{
\method{plottrace}{testpleio.1vs2}(x, xlab = "cut point", ylab = "LODdiff",
  mgp = c(1.6, 0.2, 0), ...)
}
\arguments{
\item{x}{object of class "testpleio.1vs2"}

\item{...}{parameters to be passed through to plotting functions.}
}
\description{
plot trace of \code{testpleio.1vs2} to see how \code{LODdiff}
changes as the cutting point moves.
}

