% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plottrans.LDA.R
\name{plottrans.LDA}
\alias{plottrans.LDA}
\title{LDA plot for a transband.}
\usage{
plottrans.LDA(Y, geno, nonrecomb, max.p = 100, ...)
}
\arguments{
\item{Y}{matrix of traits in the transband}

\item{geno}{genotype at the peak marker of scanone.mvn.}

\item{nonrecomb}{index for non-recombnent(in a 10cM region) individuals.}

\item{max.p}{max number of traits used for PCA analysis. This is
used to avoid rank deficiency.}
}
\description{
LDA plot for a transband.
}

