% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plottrans.LDA2.R
\name{plottrans.LDA2}
\alias{plottrans.LDA2}
\title{LDA plot for a transband.}
\usage{
plottrans.LDA2(Y, geno, nonrecomb, max.p = 100, geno2, addlegend = TRUE,
  ...)
}
\arguments{
\item{Y}{matrix of traits in the transband}

\item{geno}{genotype at the peak marker of scanone.mvn.}

\item{nonrecomb}{index for non-recombnent(in a 10cM region) individuals.}

\item{max.p}{max number of traits used for PCA analysis. This is
used to avoid rank deficiency.}
}
\description{
This version of plot color the recombinant mice by their 2 QTL genotype instead of all yellow.
}

