% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plottrans.PCA}
\alias{plottrans.PCA}
\title{PCA plot for a transband.}
\usage{
plottrans.PCA(Y, geno, nonrecomb, max.p = 100, ...)
}
\arguments{
\item{Y}{matrix of traits in the transband}

\item{geno}{genotype at the peak marker of scanone.mvn.}

\item{nonrecomb}{index for non-recombnent(in a 10cM region) individuals.}

\item{max.p}{max number of traits used for PCA analysis. This is
used to avoid rank deficiency.}
}
\description{
PCA plot for a transband.
}

