% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plottrans.tSNE.R
\name{plottrans.tSNE}
\alias{plottrans.tSNE}
\title{tsne plot for a transband.}
\usage{
plottrans.tSNE(Y, geno, nonrecomb, ...)
}
\arguments{
\item{Y}{matrix of traits in the transband}

\item{geno}{genotype at the peak marker of scanone.mvn.}

\item{nonrecomb}{index for non-recombnent(in a 10cM region) individuals.}

\item{...}{Optional graphics parameters passed to \code{\link[broman]{grayplot}}
used to avoid rank deficiency.}
}
\description{
tsne plot for a transband.
}

