% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/propn.plot.R
\name{propn.plot}
\alias{propn.plot}
\title{propotion plot.}
\usage{
propn.plot(genotype, map, pred.test, pred.score, main = "",
  weighted = FALSE)
}
\arguments{
\item{genotype}{genotype matrix, rows are samples, columns are markers.}

\item{map}{genetic map for markers.}

\item{pred.test}{vector of length n, predicted class (1, 2 or 3)
for each sample.}

\item{pred.score}{measure of how good is the prediction, when
weighted = TRUE, this score is used to order the samples.}

\item{main}{main titile of the plot.}

\item{weighted}{use pred.score as weights? default is \code{FALSE}.}
}
\value{
a proportion plot.
}
\description{
Compare pred.test to the genotype of each marker, calculate and
plot the propotion of matches at each marker position and inside
each marker interval, could be used to estimate interval for the
common eQTL in a transband.
}

