% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{scantwo.mvn}
\alias{scantwo.mvn}
\title{Multivariate normal genome 2d scan.}
\usage{
scantwo.mvn(cross, Y, chr = NULL, addcovar = NULL, intcovar = NULL,
  tol = 1e-07)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{Matrix of multiple traits with n rows.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated.  This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{addcovar}{Additive covariates.}

\item{intcovar}{Interactive covariates.}

\item{tol}{Tolerance value for the \code{qr} decomposition in
\code{lm} fitting.}
}
\description{
Genome scan with two QTL model, with allowance for possible
covariates, using multivariate normal model for the multiple
traits.
}

