% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/add.effect.R
\name{add.effect}
\alias{add.effect}
\title{find the QTL addtive and dominant effect for all traits in the transband.}
\usage{
add.effect(transbands, cross)
}
\description{
find the QTL addtive and dominant effect for all traits in the transband.
}

