% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/add.effect.R
\name{add.effect}
\alias{add.effect}
\title{find the QTL addtive and dominant effect for all traits in the transband.}
\usage{
add.effect(transbands, cross)
}
\arguments{
\item{transbands}{result from find.trans, a list of transband.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}
}
\description{
find the QTL addtive and dominant effect for all traits in the transband.
}

