% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/add.geno.R
\name{add.geno}
\alias{add.geno}
\title{find the peak of scanone.mvn and the QTL genotype.}
\usage{
add.geno(transbands, cross, regn.cM = 5, max.p = 100)
}
\arguments{
\item{regn.cM}{disctance in cM form QTL for defining non-recombinanct mice.}
}
\description{
For a list of hotspots, find the single QTL genotype and the mice
with no recombinant events in a 10cM region around the single QTL.
}

