% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/add.geno.R
\name{add.geno}
\alias{add.geno}
\title{find the peak of scanone.mvn and the QTL genotype.}
\usage{
add.geno(transbands, cross, regn.cM = 5, max.p = 100)
}
\description{
find the peak of scanone.mvn and the QTL genotype.
}

