% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/group.train.test.R
\name{group.train.test}
\alias{group.train.test}
\title{find the training and tesing set from genotypes in a given chr region.}
\usage{
group.train.test(cross, Y, chr, region.l, region.r)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{traits mapped to the region.}

\item{chr}{which chromosome is under invest.}

\item{region.l}{left bound of region under invest.}

\item{region.r}{right bound of region under invest.}
}
\value{
a list of data.train, data.test, geno.train, geno.test, map.
}
\description{
find the training and tesing set from genotypes in a given chr region.
}

