% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lm.resid.R
\name{lm.resid}
\alias{lm.resid}
\title{Calculate residual matrix.}
\usage{
lm.resid(X, Y, method = c("llt", "qr", "svd"))
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}

\item{method}{'llt' for the LLT Cholesky, 'qr' for the
column-pivoted QR decomposition, 'svd' for the Jacobi singular
value decomposition (SVD)}
}
\value{
The residual matrix.
}
\description{
Run lm(Y~X) with specific method through Rcpp and return a
residual matrix.
}

