% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_resid_cov}
\alias{lm_resid_cov}
\title{Sample covariance matrix of residual for linear model}
\usage{
lm_resid_cov(X, Y, threshold = 1e-07)
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}

\item{threshold}{Eigen decomposition is used in calculation. An
eigen value smaller than threshold * largest eigen value is
considered zero.}
}
\value{
The residual covariance matrix
}
\description{
Sample covariance matrix of residual for linear model
}

