% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_resid_cov_det}
\alias{lm_resid_cov_det}
\title{Determinant of covariance matrix of residual for linear model}
\usage{
lm_resid_cov_det(X, Y, logarithm = TRUE, threshold = 1e-07)
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}

\item{logarithm}{A bool value indicating whether determinant or
log determinate should be returned.}

\item{threshold}{Eigen decomposition is used in calculation. An
eigen value smaller than threshold * largest eigen value is
considered zero.}
}
\value{
The log determinant of residual covariance matrix
}
\description{
Determinant of covariance matrix of residual for linear model
}

