% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_resid_llt}
\alias{lm_resid_llt}
\title{Calculate the residual of lm(Y ~ X)}
\usage{
lm_resid_llt(X, Y)
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}
}
\value{
The residual matrix
}
\description{
Calculate the residual of lm(Y ~ X)
}

