% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{lm_resid_llt}
\alias{lm_resid_llt}
\title{Calculate the residual of lm(Y ~ X)}
\usage{
lm_resid_llt(X, Y)
}
\arguments{
\item{X}{A model matrix}

\item{Y}{The response matrix}
}
\value{
The residual matrix
}
\description{
Calculate the residual of lm(Y ~ X)
}

