% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_resid_svd}
\alias{lm_resid_svd}
\title{Residual of lm(Y ~ X)}
\usage{
lm_resid_svd(X, Y, threshold = 1e-07)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix}

\item{threshold}{Eigen decomposition is used in calculation. An
eigen value smaller than threshold * largest eigen value is
considered zero.}
}
\value{
The residual matrix
}
\description{
An implementation for calculate residual of linear
regression. It's ' not as fast as lm_resid_llt (about 50% slower),
but is rank revealing, so should be safe for linear correlated X.
}

