% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotGenetpattern.R
\name{plotGenetpattern}
\alias{plotGenetpattern}
\title{Plot the QTL dominance effect versus QTL additive effect.}
\usage{
plotGenetpattern(Y, genotype = NULL, cross, chr, a, d, addcovar = NULL,
  intcovar = NULL, LOD.threshold = 3, xlab = "Additive effect",
  ylab = "Dominance effect", mgp = c(1.6, 0.2, 0), ...)
}
\arguments{
\item{Y}{Matrix of multiple traits, with samples in the row,
traits in the column.}

\item{genotype}{QTL genotype for the common QTL.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{a}{vector of additive effect}

\item{d}{vector of diminance effect}

\item{addcovar}{Additive covariates, samples in the row.}

\item{intcovar}{Interactive covariates (interact with QTL genotype).}

\item{LOD.threshold}{threshold for QTL to be displayed.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{mgp}{Margin parameters}

\item{...}{Optional graphics arguments, passed to \code{\link[graphics]{plot}}}
}
\value{
a plot of QTL dominance effect versus QTL additive effect.
}
\description{
QTL additive effect is defined as a=(RR-BB)/2;
QTL dominant effect is defined as d=BR-(BB+RR)/2.
}
\examples{
data(fake.phenos)
data(listeria)
listeria <- calc.genoprob(listeria, step=1)
plotGenetpattern(Y=fake.phenos, cross=listeria, chr=1)
}

