% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotLODsign.R
\name{plotLODsign}
\alias{plotLODsign}
\title{Plot the signed LOD score versus QTL position.}
\usage{
plotLODsign(Y, cross, chr, LODsign, maxPOS, map, addcovar = NULL,
  intcovar = NULL, LOD.threshold = 3, xlab = "QTL pos (cM)",
  ylab = "signed LOD score", mgp = c(1.6, 0.2, 0), bgcolor = "gray80",
  ...)
}
\arguments{
\item{Y}{matrix, columns are quantitative traits maps to the same position.}

\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{addcovar}{Additive covariates, samples in the row.}

\item{intcovar}{Interactive covariates (interact with QTL genotype).}

\item{LOD.threshold}{threshold for QTL to be displayed.}

\item{...}{Optional graphics arguments}
}
\value{
a plot the signed LOD score versus QTL position for multiple traits.
}
\description{
For each trait, the QTL effect direction is used as the sign of
the LOD score.
}
\examples{
data(fake.phenos)
data(listeria)
listeria <- calc.genoprob(listeria, step=1)
chr <- 1
plotLODsign(Y, listeria, chr)
}

