% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plottrace.R
\name{plottrace}
\alias{plottrace}
\title{plot trace}
\usage{
plottrace(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{parameters to be passed through to plotting functions.}
}
\description{
plot trace of a searching method to see how the object function
changes as the search goes.
}

