\name{plottrace.testpleio.1vs2}
\alias{plottrace.testpleio.1vs2}
\title{plot trace}
\usage{
\method{plottrace}{testpleio.1vs2}(x)
}
\arguments{
  \item{x}{object of class "testpleio.1vs2"}
}
\description{
plot trace of \code{testpleio.1vs2} to see how
\code{LODdiff} changes as the cutting point moves.
}

