% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plottrans.PCA.sex}
\alias{plottrans.PCA.sex}
\title{PCA plot for a transband, color by sex.}
\usage{
plottrans.PCA.sex(Y, sex, max.p = 100, ...)
}
\arguments{
\item{Y}{matrix of traits in the transband}

\item{sex}{Female =1, Male=2}

\item{max.p}{max number of traits used for PCA analysis. This is
used to avoid rank deficiency.}
}
\description{
PCA plot for a transband, color by sex.
}

