% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/scantwo.mvn.R
\name{scantwo.mvn}
\alias{scantwo.mvn}
\title{Multivariate normal genome 2d scan.}
\usage{
scantwo.mvn(cross, Y, chr = NULL, addcovar = NULL, intcovar = NULL,
  method = c("maxlikelihood", "pillaitrace"), tol = 1e-07)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{Matrix of multiple traits, with samples in the row,
traits in the column.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{addcovar}{Additive covariates, samples in the row.}

\item{intcovar}{Interactive covariates (interact with QTL genotype).}

\item{method}{choose from "ML" for maximum likelihood method for
LOD score, "Pillai", "Wilks", "Hotelling-Lawley", or "Roy" for the
corresponding MANOVA methods.}

\item{tol}{Tolerance value for the \code{qr} decomposition in
\code{lm} fitting.}
}
\description{
Genome scan with two QTL model, with allowance for possible
covariates, using multivariate normal model for the multiple
traits.
}

