% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/testpleio.1vsp.R
\name{testpleio.1vsp}
\alias{testpleio.1vsp}
\title{Test of pleiotrophic QTL vs close linked QTLs.}
\usage{
testpleio.1vsp(cross, Y, chr = "6", addcovar = NULL, intcovar = NULL,
  n.simu = 1000, tol = 1e-07)
}
\arguments{
\item{cross}{An object of class \code{cross}. See
\code{read.cross} for details.}

\item{Y}{Matrix of multiple traits, with samples in the row,
traits in the column.}

\item{chr}{Optional vector indicating the chromosomes for which
LOD scores should be calculated. This should be a vector of
character strings referring to chromosomes by name; numeric values
are converted to strings.}

\item{addcovar}{Additive covariates, samples in the row.}

\item{intcovar}{Interactive covariates (interact with QTL genotype).}

\item{n.simu}{Number of simulations for p-value.}

\item{tol}{Tolerance value for the \code{qr} decomposition in
\code{lm} fitting.}
}
\value{
a list of LOD1, LODp, LODdiff, LOD1.pos and maxPOS ... P-value
\item{LOD1}{LOD score of one dimensional joint mapping.}
\item{LODp}{LOD score for estimate of p QTL model, each trait
            is influenced by a separate QTL.}
\item{LODdiff}{Difference of best one QTL model and best p QTL model,
               LODdiff = max(LODp) - max(LOD1)}
\item{pvalue}{P-value from parametric bootstrap simulations.}
}
\description{
Test if multitriats are controlled by one pleiotrophic QTL or
multiple close linked QTLs, allowing each trait to have its own
QTL. p-value is calculated from parametric bootstrap, which takes
the estimated parameters under null hypothesis and generate data
from them to get the null empirical distribution of test
statistic.
}
\examples{
data(fake.phenos)
data(listeria)
listeria <- calc.genoprob(listeria, step=1)
obj <- testpleio.1vsp(cross=listeria, Y=fake.phenos, chr=1, n.simu=100)
summary(obj)
plot(obj)
}

